Add Number class and refactor sum and divide functions
This commit introduces a new class called "Number" in the "number.py" module under the "app/utils" directory. The "sum" and "divide" functions have been refactored to methods within the "Number" class. This change improves the organization and encapsulation of the code.

The purpose of this commit is to enhance the code structure by introducing a class for number operations and encapsulating the related functions within it.